chore: generate TS types for DHIS2 schemas from OpenApi specs
These types are auto-generated so they are gitignored.
The types are automatically generated during postinstall,
which ensures they are present.
Updating the types can be done by calling `yarn generate-types`
The generation script only creates types for the DHIS2 schemas,
so the app has acces to the data model. It is also possible to
generate more output for endpoints and query parameters, but
I am not sure to what extend this is useful or accurate.
The script generates and format the `.ts` files but they
are excluded from linting/formatting because they are gitignored.