chore: manually create import aliases for Vite and Vitest
Note that these are exported from `vitest.config.ts` which is not
the most logical option. I opted for this because with our current
TS config it is not possible to import from a `.mts` file:
```
An import path can only end with a '.mts' extension when
'allowImportingTsExtensions' is enabled.
```
I preferred exporting from an unusual location over changing the
TS config