feat: redux store for visualization ui config (#48)
Implements part of DHIS2-19955

Description
This PR sets up the redux store needed to support displaying the visualization in the UI. Since this state is about how the visualization is displayed in the app, I named it "vis-ui-config". Open for other suggestions though.

The getVisualizationUiConfig function is the same as what we have in LL, but with some renaming and simplification.