feat: layout chip (#49)
Implements part of DHIS2-20034

Description
Implement the Chip component that goes in the visualization layout.
The implementation doesn't include showing the count of conditions (to prevent this PR from getting too big).

The logic in getChipItemText is pretty much unchanged from LL, except for clean-up and adding types.

The component architecture of separating Chip and ChipBase components is kept from LL,
in anticipation of upcoming Dnd implementation. (dnd uses ChipBase for dragging)