chore: rename InputType to OutputType (#78)
Implements DHIS2-20327

Description
We originally had decided that outputType would be renamed to inputType and that would be the term used throughout the code. But it turns out that there are several places where the original vis (with outputType prop) are fetched and so that renaming became cumbersome. When we also found out that the new sidebar design would no longer even use the term "Input type", then it became clear we should just drop the renaming and use outputType throughout the code.