refactor: remove dryRun, reworked helper functions
The validation is per vis type, since the rules can be different.
The validation functions throw errors which will eventually be used to
control the error screen component in the canvas area.
A helper function is running the validate function and returns a
boolean, this is used to enable/disable the Download menu.